chore: fix folder explorer opening issue
