+++ src/App.tsx
@@ -14,7 +14,8 @@ import {Project} from '@models/appconfig';
 import {Size} from '@models/window';
 
 import {useAppSelector} from '@redux/hooks';
-import {setOpenProject} from '@redux/reducers/appConfig';
+import {setCreateProject, setOpenProject} from '@redux/reducers/appConfig';
+import {closeFolderExplorer} from '@redux/reducers/ui';
 import {kubeConfigContextSelector, kubeConfigPathSelector, settingsSelector} from '@redux/selectors';
 import {loadContexts} from '@redux/thunks/loadKubeConfig';
 
@@ -38,9 +39,12 @@ import {
   UpdateModal,
 } from '@organisms';
 
+import FileExplorer from '@components/atoms/FileExplorer';
 import ChangeFiltersConfirmModal from '@components/molecules/ChangeFiltersConfirmModal/ChangeFiltersConfirmModal';
 import SaveResourceToFileFolderModal from '@components/molecules/SaveResourcesToFileFolderModal';
 import {CreateProjectModal} from '@components/organisms/CreateProjectModal';
+
+import {useFileExplorer} from '@hooks/useFileExplorer';
 
 import {getFileStats} from '@utils/files';
 import {useWindowSize} from '@utils/hooks';
@@ -108,6 +112,25 @@ const App = () => {
     [kubeConfigPath, dispatch, kubeConfigContext, rootFile]
   );
 
+  const isFolderExplorerOpen = useAppSelector(state => state.ui.folderExplorer.isOpen);
+
+  const {openFileExplorer, fileExplorerProps} = useFileExplorer(
+    ({folderPath}) => {
+      if (folderPath) {
+        dispatch(setCreateProject({rootFolder: folderPath}));
+      }
+    },
+    {isDirectoryExplorer: true}
+  );
+
+  useEffect(() => {
+    if (isFolderExplorerOpen) {
+      openFileExplorer();
+      dispatch(closeFolderExplorer());
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [isFolderExplorerOpen]);
+
   return (
     <AppContext.Provider value={{windowSize: size}}>
       <AppContainer $height={size.height} $width={size.width}>
@@ -119,6 +142,7 @@ const App = () => {
           <PaneManager />
           <PageFooter />
         </MainContainer>
+        <FileExplorer {...fileExplorerProps} />
         <LocalResourceDiffModal />
         <ClusterResourceDiffModal />
         <StartupModal />
